Improve streaming of submitted files.
Add a new (optional) parameter to submit() called 'openfile' that can
take an open file handle which will be read when submitting the file to
assemblyline. This is useful in the cases where the file being uploaded
was created using something like TemporaryFile or SpooledTemporaryFile,
where there might be no directory entry for the file.

In addition, use MultipartEncoder from requests_toolbelt so that
'requests' doesn't read the entire file in to memory before POSTing it
to assemblyline.

Make sure to clean up the temporary file that was created if 'content'
was given to submit().